'Refactored by Sourcery'

+++ move_files_over_x_days.py
@@ -41,9 +41,7 @@ args_parser.add_argument(
 )
 args = args_parser.parse_args()
 
-if args.days < 0:
-    args.days = 0
-
+args.days = max(args.days, 0)
 src = args.src  # Set the source directory
 dst = args.dst  # Set the destination directory
 days = args.days  # Set the number of days
@@ -53,8 +51,5 @@ if not os.path.exists(dst):
     os.mkdir(dst)
 
 for f in os.listdir(src):  # Loop through all the files in the source directory
-    if (
-        os.stat(f).st_mtime < now - days * 86400
-    ):  # Work out how old they are, if they are older than 240 days old
-        if os.path.isfile(f):  # Check it's a file
-            shutil.move(f, dst)  # Move the files
+    if (os.stat(f).st_mtime < now - days * 86400) and os.path.isfile(f):
+        shutil.move(f, dst)  # Move the files
